+++ secrets/secrets.nix
@@ -18,4 +18,9 @@ in {
     inherit (users) enzime_phi;
     inherit (hosts) phi;
   };
+
+  "zshrc_phi.age".publicKeys = builtins.attrValues {
+    inherit (users) enzime_phi;
+    inherit (hosts) phi;
+  };
 }